requirements.txt
aiomysql==0.0.21
alembic==1.4.3
arrow==0.17.0
binaryornot==0.4.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cookiecutter==1.7.2
fastapi==0.70.0
h11==0.12.0
idna==2.10
Jinja2==2.11.2
jinja2-time==0.2.0
Mako==1.1.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
poyo==0.5.0
protobuf==3.14.0
pydantic>=1.8.0,<2.0.0
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.16.0
text-unidecode==1.3
urllib3==1.26.2
uvicorn>=0.15.0,<0.16.0
python-dotenv==1.0.0
python-keycloak>=2.6.0
